backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
channels==4.1.0
requests==2.31.0
Pillow==10.3.0

# ML
torch==2.1.0
transformers==4.40.0
numpy==1.26.4
scikit-learn==1.4.2
pandas==2.2.2
datasets==2.19.1
accelerate==0.30.0
